changed int to Long in ExtendedDiscountViews

+++ src/main/java/com/exadel/discountwebapp/statistics/dto/extendeddto/ExtendedDiscountViews.java
@@ -16,5 +16,5 @@ public interface ExtendedDiscountViews {
     String getVendorTitle();
     String getCategoryTitle();
     Long getViewNumber();
-    int getActivated();
+    Long getActivated();
 }
